pyproject.toml
[tool.poetry]
name = "Buku CRUD API"
version = "0.1.0"
description = "App Description"
authors = ["Adrian Maulani <maulani.ad@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
drf-standardized-errors = "^0.14.1"
psycopg = "^3.2.1"
psycopg-binary = "^3.2.1"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
django-stubs = "^5.0.4"
djangorestframework-stubs = "^3.15.0"
rich = "^13.7.1"
ruff = "^0.5.5"

[tool.ruff]
line-length = 120

[tool.ruff.format]
indent-style = "space"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
